Refactor to ResolutionExtensions
